==> 0.1-Champ_Preprocessing.ipynb <==
# Data Preprocessing #

### Imports ###

import pandas as pd
import numpy as np
import datetime

### Data Preview ###

data = pd.read_csv('Animal_IO_Data_0.csv') # https://data.louisvilleky.gov/sites/default/files/Animal_IO_Data_0.csv
print("Total Observations: ", len(data))
data.head(10)

print(data.columns, '\n')

### Scope ###
The scope of the project was limited to only cats and dogs

data = data[(data['AnimalType'] == 'CAT') | (data['AnimalType'] == 'DOG')]
print("Total Observations: ", len(data))

### Dropping Unhelpful Rows and Columns ###

#### For the purposes of this project, OutsideAsilomarStatus is redundant (OutcomeType) and ReproductiveStatusAtOutcome is irrelevant. ####

data = data.drop(['OutcomeAsilomarStatus', 'ReproductiveStatusAtOutcome'], axis = 1)

#### IntakeAsilomarStatus is redundant because the same IntakeInternalStatus always has the same IntakeAsilomarStatus. ####

data = data.drop(['IntakeAsilomarStatus'], axis = 1)

#### Rows without an OutcomeType will not be helpful in training our model. ####

data = data[pd.notnull(data['OutcomeType'])]
data = data.reset_index(drop = True)

#### Certain IntakeTypes will prevent any possible intervention in the outcome and certain outcomes cannot be avoided. ####
For example, if an animal is brought in through a euthanasia request, nothing can be done to change the outcome. There are several instances of this and so they will not be included.

NPI_IntakeTypes = ['DISPOSAL', # disposal of dead animal
                   'ET REQUEST',  # euthanasia request
                   'EUTH REQ', # euthanasia request
                   'OUTSURGERY', # surgery
                   'QUARANTINE']  # isolated due to health/aggression reasons (almost always euthanized with fair reason)

NPI_IntakeSubtypes = ['BITE', # almost always euthanized for aggression
                      'EUTH REQ',  # euthanasia request
                      'DANGER DOG',  # dangerous dog (almost always euthanized with fair reason)
                      'POST SURG', # surgery
                      'RESCUE GRP', # always transferred
                      'POTDANGER']  # potential danger (almost always euthanized with fair reason)

NPI_OutcomeTypes = ['RTO',  # returned to owner (lost and found)
                    'DISPOSAL',  # disposing of dead animal
                    'TNR', # trap and release (animals taken and released after spayed/neutered)
                    'SNR', # similar to TNR
                    'RTF', # similar to TNR
                    'MISSING',  # uncontrollable/unknown outcome
                    'MISSING EX',  # uncontrollable/unknown outcome
                    'LOST EXP',  # uncontrollable/unknown outcome
                    'NO SHOW',  # uncontrollable/unknown outcome
                    'RELEASE',  # likely similar to TNR
                    'RELEASED']  # likely similar to TNR

NPI_ITs = data[data['IntakeType'].isin(NPI_IntakeTypes)].index
data = data.drop(NPI_ITs)

NPI_ISs = data[data['IntakeSubtype'].isin(NPI_IntakeSubtypes)].index
data = data.drop(NPI_ISs)

NPI_OTs = data[data['OutcomeType'].isin(NPI_OutcomeTypes)].index
data = data.drop(NPI_OTs)

data = data[data['IntakeInternalStatus'] != 'DEAD']  # animals that were dead on arrival
data = data[data['OutcomeSubtype'] != 'REQUESTED'] # euthanasia requested (improperly documented in IntakeType)

data = data.reset_index(drop=True)

### Dropping Columns with Many Missing Values ###

#### Checking for Missingness ####

data.isna().sum()[data.isna().sum() > 0]/len(data)

#### Dropping columns with a very high percentage of missing values ####

data = data.drop(
    ['SecondaryBreed', 'SecondaryColor', 'IntakeReason', 'OutcomeReason', 'OutcomeInternalStatus'], 
    axis = 1)

Note: Despite having a high percentage of missing values, DOB was not dropped because of its high relevance.

### Dealing with Duplicate Entries ####

#### Checking for duplicate AnimalIDs ####

dupl_IDs = pd.concat(i for _, i in data.groupby("AnimalID") if len(i) > 1)
dupl_IDs = dupl_IDs.sort_values(by = ['AnimalID', 'IntakeDate'])
print("Number of duplicate AnimalIDs: ", int(len(dupl_IDs)/2))

dupl_IDs.head(10)

#### Dropping all but the first observation when observations are an EXACT match and reseting index ####

data = data.drop_duplicates()

#### Dropping latter of two observations when the same AnimalID results in 2 'EUTH'/'DIED' outcomes

died_IDs = dupl_IDs[(dupl_IDs['OutcomeType'] == 'EUTH') | (dupl_IDs['OutcomeType'] == 'DIED') | (dupl_IDs['OutcomeType'] == 
                                                                                                'ET PROCESS')]
dupl_died_IDs = pd.concat(i for _, i in died_IDs.groupby("AnimalID") if len(i) > 1)
dupl_TBR = dupl_died_IDs.iloc[range(1, len(dupl_died_IDs), 2)]

data = data.drop(dupl_TBR.index)
data = data.reset_index(drop = True)

#### Rechecking number of duplicate AnimalIDs ####

dupl_IDs = pd.concat(i for _, i in data.groupby("AnimalID") if len(i) > 1)
dupl_IDs = dupl_IDs.sort_values(by = ['AnimalID', 'IntakeDate'])
print("Number of remaining duplicate AnimalIDs: ", int(len(dupl_IDs)/2))

#### Removing returned animals from duplicate subset ####
Returned animals count as two observations since being returned often gave a different outcome than the initial outcome. Because of this, they should not be included in the dupl_IDs subset. Removal from this set does not have any affect on the main dataset.

RETURN_IDs = dupl_IDs[(dupl_IDs['IntakeType'] == 'RETURN') | (dupl_IDs['IntakeSubtype'] == 'RETURN')]['AnimalID'].values
dupl_IDs = dupl_IDs.drop(dupl_IDs[dupl_IDs['AnimalID'].isin(RETURN_IDs)].index)

print("Number of actual remaining duplicate AnimalIDs: ", int(len(dupl_IDs)/2))

#### Examining remaining duplicate AnimalIDs ####

dupl_IDs.head(50)

It appears that the remaining duplicate AnimalIDs do belong do the same animals. However, these observations should still be useful because they take place under different circumstances and often have different outcomes.

### Basic Feature Modifications ###

A few feature modifications were done that will apply to both the dog and cat models.

#### Changing Genders 'NEUTERED MALE' and 'SPAYED FEMALE' to simply 'MALE' and 'FEMALE' ####
The labels 'neutered' and 'spayed' should correspond to 'fertile' and 'altered' for ReproductiveStatusAtIntake, but they do not. It was determined that due to the frequency of each value, the 'fertile' and 'altered' labels were true, whereas the 'neutered' and 'spayed' labels were incorrectly done.

def gender_transform(gender):

    if gender == 'NEUTERED MALE':
        return 'MALE'
    elif gender == 'SPAYED FEMALE':
        return 'FEMALE'
    else:
        return gender
    
data['Gender'] = data['Gender'].apply(gender_transform)

#### Changing OutcomeType 'ET PROCESS' to 'EUTH' ####
'ET PROCESS' is used to denote an animal that is going to be euthanized. However, this outcome type was never changed when they were euthanized.

def ETtoEUTH(OutcomeType):
    if OutcomeType == 'ET PROCESS':
        return 'EUTH'
    else:
        return OutcomeType
    
data['OutcomeType'] = data['OutcomeType'].apply(ETtoEUTH)

#### Changing OutcomeSubtypes 'TIME' and 'SPACE' to 'TIME/SPACE' ####
'TIME/SPACE' should cover 'TIME' and 'SPACE' and the differentiation is likely arbitrary.

def TStoTS(OutcomeSubtype):
    if ((OutcomeSubtype == 'TIME') | (OutcomeSubtype == 'SPACE')):
        return 'TIME/SPACE'
    else:
        return OutcomeSubtype
    
data['OutcomeSubtype'] = data['OutcomeSubtype'].apply(TStoTS)

#### Getting value from dates ####
Date columns were changed to a datetime type and were then subsequently used to creature other features, such as Intake/Outcome Month, Intake/Outcome Year, Intake/Outcome DotW (day of the week), IntakeAge, and TimeUntilOutcome.

IntakeDate = pd.to_datetime(data['IntakeDate'], infer_datetime_format = True).dt.strftime('%m-%d-%Y')
OutcomeDate = pd.to_datetime(data['OutcomeDate'], infer_datetime_format = True).dt.strftime('%m-%d-%Y')
DOB = pd.to_datetime(data['DOB'], infer_datetime_format = True).dt.strftime('%m-%d-%Y')

data['IntakeDate'] = pd.to_datetime(IntakeDate)
data['OutcomeDate'] = pd.to_datetime(OutcomeDate)
data['DOB'] = pd.to_datetime(DOB)

def get_month(date):
    return date.month

def get_year(date):
    return date.year

def get_dotw(date):
    return date.dayofweek


data['IntakeMonth'] = data['IntakeDate'].apply(get_month)
data['IntakeYear'] = data['IntakeDate'].apply(get_year)
data['IntakeDotW'] = data['IntakeDate'].apply(get_dotw)

data = data[-data['OutcomeDate'].isna()]
data['OutcomeMonth'] = data['OutcomeDate'].apply(get_month)
data['OutcomeYear'] = data['OutcomeDate'].apply(get_year)
data['OutcomeDotW'] = data['OutcomeDate'].apply(get_dotw)

IntakeAge = data['IntakeDate'] - data['DOB']
TimeUntilOutcome = data['OutcomeDate'] - data['IntakeDate']

data['IntakeAge'] = pd.to_numeric(IntakeAge.dt.days, downcast='integer')

def invalid_age(IntakeAge):
    if IntakeAge < 0:
        return np.nan
    else:
        return IntakeAge

data['IntakeAge'] = data['IntakeAge'].apply(invalid_age) # replacing negative ages with NaN

data['TimeUntilOutcome'] = pd.to_numeric(TimeUntilOutcome.dt.days, downcast='integer')
data = data[data['TimeUntilOutcome'] >= 0]  # removing data where OutcomeDate was before IntakeDate

#### Adding 'LiveOutcome' column ####
The model seeks to classify whether an animal is euthanized or not, so the binary 'LiveOutcome' feature is created.

def live_outcome(OutcomeType):
    if ((OutcomeType == 'EUTH') | (OutcomeType == 'DIED')):
        return 0
    else:
        return 1
    
data['LiveOutcome'] = data['OutcomeType'].apply(live_outcome)

#### Removing newly redundant columns ####

data = data.drop(['IntakeDate', 'OutcomeDate', 'DOB'], axis=1)

### Preprocessed Data Preview ###

print("Total Observations: ", len(data))
data.head(20)

### Saving Preprocessed Data ###

data.to_csv('preprocessed_data.csv')
