==> 1.1-Champ_CAT_Preprocessing.ipynb <==
# CATS: Further Data Preprocessing #

### Imports ###

import pandas as pd
import numpy as np
import datetime
import re

### Data Preview ###
Cat data was extracted from the data exported from the Champ_Preprocessing notebook.

data = pd.read_csv('preprocessed_data.csv', index_col=0)

cat_data = data[data['AnimalType'] == 'CAT'].reset_index(drop=True)
cat_data = cat_data.drop('AnimalType', axis=1)
print("Total Observations: ", len(cat_data))
cat_data.head()

### Feature Value Reduction ###
The number of feature values  were reduced to help with training the model later on and to categorize values that may have been assigned differently arbitrarily.

#### Combining IntakeType and IntakeSubtype ####

cat_data['IntakeTypes'] = cat_data['IntakeType'].map(str) + ',' + cat_data['IntakeSubtype']
cat_data = cat_data.drop(['IntakeType', 'IntakeSubtype'], axis=1)

#### Classifying the IntakeTypes into 10 categories ####
Many of the combined intake types were very similar and can be grouped together.

def cat_IntakeType(IntakeTypes):
    IntakeTypes = str(IntakeTypes)
    t1 = re.compile("CONFISCATE,ABANDONED|CONFISCATE,NEGLECT|CONFISCATE,CRUELTY")
    t2 = re.compile(
        "CONFISCATE,EVICTION|CONFISCATE,UNPERMITED|CONFISCATE,POLICE|CONFISCATE,HOSPITAL|CONFISCATE,OWNER DIED|CONFISCATE,RESTRAINT")
    t3 = re.compile("FOSTER,RETURN|RETURN,FOSTER")
    t4 = re.compile("OWNER SUR,FIELD|KHS,OWNER SUR")
    t5 = re.compile("OWNER SUR,OTC")
    t6 = re.compile("OWNER SUR,RETURN 30|RETURN,ADOPTION|RETURN,FIELD|RETURN,K HUMANE S|RETURN,OTC")
    t7 = re.compile("KHS,STRAY")
    t8 = re.compile("STRAY,FIELD|STRAY,STRAY")
    t9 = re.compile("STRAY,OTC|OWNER SUR,STRAY")
    
    if t1.match(IntakeTypes):
        return 'CONFISCATE,CRUELTY'
    elif t2.match(IntakeTypes):
        return 'CONFISCATE,OTHER'
    elif t3.match(IntakeTypes):
        return 'FOSTER'
    elif t4.match(IntakeTypes):
        return 'OWNERSUR,FIELD'
    elif t5.match(IntakeTypes):
        return 'OWNERSUR,OTC'
    elif t6.match(IntakeTypes):
        return 'RETURN'
    elif t7.match(IntakeTypes):
        return 'STRAY,KHS'
    elif t8.match(IntakeTypes):
        return 'STRAY,FIELD'
    elif t9.match(IntakeTypes):
        return 'STRAY,OTC'
    else:
        return 'OTHER'

cat_data['IntakeTypes'] = cat_data['IntakeTypes'].apply(cat_IntakeType)
cat_data.groupby('IntakeTypes').count()['AnimalID']

#### Categorizing cat colors ####

def cat_color_class(color):
    color = str(color)
    c1 = re.compile("ORANGE|APRICOT|FLAME|RED|TABBY")
    c2 = re.compile("GRAY|BLUE|SEAL|SILVER")
    c3 = re.compile("CREAM|BUFF|GOLD|TAN|YELLOW")
    c4 = re.compile("BROWN|CHOCOLATE|MAHOGANY|SABLE")
    c5 = re.compile("WHITE|LYNX")
    c6 = re.compile("TORTIE|TORTI")
    c7 = re.compile("BLACK")
    c8 = re.compile("CALICO")
   
    if c1.match(color):
        return 'ORANGE'
    elif c2.match(color):
        return 'GRAY'
    elif c3.match(color):
        return 'CREAM'
    elif c4.match(color):
        return 'BROWN'
    elif c5.match(color):
        return 'WHITE'
    elif c6.match(color):
        return 'TORTIE'
    elif c7.match(color):
        return 'BLACK'
    elif c8.match(color):
        return 'CALICO'
    else:
        return 'OTHER'

cat_data['Color'] = cat_data['PrimaryColor'].str.split().str[0]
cat_data['Color'] = cat_data['Color'].apply(cat_color_class)
cat_data = cat_data.drop('PrimaryColor', axis=1)

cat_data.groupby('Color').count()['AnimalID']

#### Categorizing cat breeds ####

def cat_breed(breed):
    breed = str(breed)
    
    if (breed == 'DOMESTIC LONGHAIR'):
        return 'LONG'
    elif (breed == 'DOMESTIC MEDIUMHAIR'):
        return 'MED'
    elif (breed == 'DOMESTIC SHORTHAIR'):
        return 'SHORT'
    elif (breed == 'AMERICAN SHORTHAIR'):
        return 'AMSHORT'
    elif (breed == 'SIAMESE'):
        return 'SIAMESE'
    elif (breed != 'NaN'):
        return 'OTHER'
    else:
        return 'NONE'

cat_data['Breed'] = cat_data['PrimaryBreed'].apply(cat_breed)
cat_data = cat_data.drop('PrimaryBreed', axis=1)

cat_data.groupby('Breed').count()['AnimalID']

### Preprocessed Cat Data Preview ###

print("Total Observations: ", len(cat_data))
cat_data.head(20)

### Saving Preprocessed Cat Data ###

cat_data.to_csv('cat_data.csv')
