==> 1.3-Champ_CAT_Transformation.ipynb <==
# CATS: Data Transformation #

### Imports ###

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

### Data Preview ###
Cat data was extracted from the data exported from the Champ_CAT_Preprocessing notebook.

all_cat_data = pd.read_csv('cat_data.csv', index_col=0)

print("Total entries: ", len(all_cat_data))
all_cat_data.head()

### Dropping Exploratory Features ###

cat_data = all_cat_data.drop(['AnimalID', 'OutcomeType', 'OutcomeSubtype', 'OutcomeYear', 'OutcomeDotW', 'OutcomeMonth',
                              'TimeUntilOutcome'], axis = 1)

### Preparing Cat Data for Transformation ###

#### Changing Integer Types to Object Types ####

cat_data = cat_data.astype({
        'IntakeMonth': 'object',
        'IntakeYear': 'object',
        'IntakeDotW': 'object',
        'LiveOutcome': 'object'})

#### Bucketing IntakeAge ####

cat_data['IntakeAge'] = pd.cut(cat_data['IntakeAge'], [0,90,180,270,360,720,1080,2160,3240])

### Transforming Cat Data with Dummies ###

cat_transf = pd.get_dummies(cat_data, drop_first = True)
cat_transf.head()

### Dropping Highly Correlated Features ###

corr = cat_transf.corr()

high_val = 0.5

high_corrs = corr[(corr > high_val) & (corr < 1)].sum()
high_corrs[high_corrs > 0]

cat_transf = cat_transf.drop('ReproductiveStatusAtIntake_UNKNOWN', axis = 1)

### Transformed Cat Data Preview ###

cat_transf.head(20)

### Saving Transformed Cat Data ###

cat_transf.to_csv('cat_transformed_data.csv')
