==> 2.3-Champ_DOG_Transformation.ipynb <==
# DOGS: Data Transformation #

### Imports ###

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

### Data Preview ###
Dog data was extracted from the data exported from the Champ_DOG_Preprocessing notebook.

all_dog_data = pd.read_csv('dog_data.csv', index_col=0)

print("Total entries: ", len(all_dog_data))
all_dog_data.head()

### Dropping Exploratory Features ###

dog_data = all_dog_data.drop(['AnimalID', 'OutcomeType', 'OutcomeSubtype', 'OutcomeYear', 'OutcomeDotW', 'OutcomeMonth',
                              'TimeUntilOutcome', 'PrimaryBreed'], axis = 1)

### Preparing Dog Data for Transformation ###

#### Changing Integer Types to Object Types ####

dog_data = dog_data.astype({
        'IntakeMonth': 'object',
        'IntakeYear': 'object',
        'IntakeDotW': 'object',
        'LiveOutcome': 'object'})

#### Bucketing IntakeAge ####

dog_data['IntakeAge'] = pd.cut(dog_data['IntakeAge'], [0,90,180,270,360,720,1080,2160,3240])

### Transforming Dog Data with Dummies ###

dog_transf = pd.get_dummies(dog_data, drop_first = True)
dog_transf.head()

### Dropping Highly Correlated Features ###

corr = dog_transf.corr()

high_val = 0.5

high_corrs = corr[(corr > high_val) & (corr < 1)].sum()
high_corrs[high_corrs > 0]

dog_transf = dog_transf.drop('ReproductiveStatusAtIntake_UNKNOWN', axis = 1)

### Transformed Dog Data Preview ###

dog_transf.head(20)

### Saving Transformed Dog Data ###

dog_transf.to_csv('dog_transformed_data.csv')
